Move splitter function to internal
